Handle updates with a system of Handlers
The GildedRose has UpdateHandlers that handle the different behaviours when updating.
Those UpdateHandlers also have a list of ExpirationUpdateHandlers that handle what happens when an item is expired.

Initially I had the GildedRose hold both the Expiration and UpdateHandlers, but since the behaviour when dealing with expiration also includes updating the quality, these were taken together.